+++ app/src/main/java/com/yoyiyi/soleil/bean/app/video/VideoDetailComment.java
@@ -91,7 +91,7 @@ public class VideoDetailComment {
             public int floor;
             public int state;
             public int attr;
-            public int ctime;
+            public long ctime;
             public String rpid_str;
             public String root_str;
             public String parent_str;
@@ -261,7 +261,7 @@ public class VideoDetailComment {
             public int floor;
             public int state;
             public int attr;
-            public int ctime;
+            public long ctime;
             public String rpid_str;
             public String root_str;
             public String parent_str;
